+++ src/test/java/com/ibm/graph/client/schema/EntityIndexTests.java
@@ -12,13 +12,18 @@ public class EntityIndexTests {
 	@Ignore("tests are included in subclass tests")
 	@Test
 	public void testEntityIndex() throws Exception {
-	    // tests are included in subclass tests
+	    // tests are included in subclass tests:
+	    //  com.ibm.graph.client.schema.VertexIndex
+	    //  com.ibm.graph.client.schema.EdgeIndex
 	}
 
 	@Ignore("tests are included in subclass tests")
 	@Test
 	public void testEntityIndexErrorHandling() throws Exception {
 	    // tests are included in subclass tests
+	    //  com.ibm.graph.client.schema.VertexIndex
+	    //  com.ibm.graph.client.schema.EdgeIndex
+
 	}
 }
